♻️(front) migrate UploadField tests to @testing-library
We are migrating all our tests from enzyme to @testing-library. We also
decide to rename the module containing the component in `index.tsx`.
