+++ src/frontend/components/UploadField/index.spec.tsx
@@ -1,0 +1,31 @@
+import { cleanup, fireEvent, render, wait } from '@testing-library/react';
+import React from 'react';
+
+import { UploadField } from '.';
+
+describe('<UploadField />', () => {
+  afterEach(cleanup);
+
+  it('renders a Dropzone with the relevant messages', () => {
+    const { getByText } = render(<UploadField onContentUpdated={jest.fn()} />);
+    getByText('Select a file to upload');
+  });
+
+  it('passes the file to the callback', async () => {
+    const callback = jest.fn();
+    const { container, getByText } = render(
+      <UploadField onContentUpdated={callback} />,
+    );
+
+    const file = new File(['(⌐□_□)'], 'course.mp4', { type: 'video/mp4' });
+
+    fireEvent.change(container.querySelector('input[type="file"]')!, {
+      target: {
+        files: [file],
+      },
+    });
+    await wait();
+
+    expect(callback).toHaveBeenCalledWith(file);
+  });
+});
